Update: 完成文件夹创建功能

+++ xss_receiver/controllers/upload_file_controller.py
@@ -1,6 +1,7 @@
 import asyncio
-from os import listdir, unlink, rename
-from os.path import join, exists, getsize, getmtime, isdir
+from os import scandir, unlink, rename, mkdir
+from os.path import join, exists, getsize, getmtime, isdir, isfile
+import shutil
 
 import sanic
 from sanic import Blueprint, json
@@ -9,7 +10,7 @@ from werkzeug.utils import secure_filename
 from xss_receiver import system_config
 from xss_receiver.jwt_auth import auth_required
 from xss_receiver.response import Response
-from xss_receiver.utils import write_file, read_file
+from xss_receiver.utils import write_file, read_file, secure_filename_with_directory
 
 upload_file_controller = Blueprint('upload_file_controller', __name__)
 
@@ -19,7 +20,7 @@ upload_file_controller = Blueprint('upload_file_controller', __name__)
 async def add(request: sanic.Request):
     file = request.files.get('file', None)
     if file:
-        filename = secure_filename(file.name)
+        filename = secure_filename_with_directory(file.name)
         path = join(system_config.UPLOAD_PATH, filename)
         if not exists(path):
             asyncio.create_task(write_file(path, file.body))
@@ -33,15 +34,24 @@ async def add(request: sanic.Request):
 @upload_file_controller.route('/list', methods=['GET'])
 @auth_required
 async def file_list(request: sanic.Request):
-    files = listdir(system_config.UPLOAD_PATH)
-    payload = []
+    def get_dir_files(path: str, base=''):
+        files = []
+        for entry in scandir(path):
+            files.append({
+                'filename': entry.name,
+                'path': join(base, entry.name),
+                'size': entry.stat().st_size,
+                'mttime': entry.stat().st_mtime,
+                'dir': entry.is_dir()
+            })
+        files = sorted(files, key=lambda x: x['filename'])
+        return files
 
-    for filename in files:
-        path = join(system_config.UPLOAD_PATH, filename)
-        payload.append({'filename': filename, 'size': getsize(path), 'mttime': getmtime(path), 'dir': isdir(path)})
-
-    payload = sorted(payload, key=lambda x: x['filename'])
-    return json(Response.success('', payload))
+    files = get_dir_files(system_config.UPLOAD_PATH)
+    for file in files:  # 只支持一层深度
+        if file['dir']:
+            file['children'] = get_dir_files(join(system_config.UPLOAD_PATH, file['filename']), file['filename'])
+    return json(Response.success('', files))
 
 
 @upload_file_controller.route('/delete', methods=['POST'])
@@ -50,9 +60,9 @@ async def delete(request: sanic.Request):
     if isinstance(request.json, dict):
         filename = request.json.get('filename', None)
         if isinstance(filename, str):
-            filename = secure_filename(filename)
+            filename = secure_filename_with_directory(filename)
             path = join(system_config.UPLOAD_PATH, filename)
-            if exists(path):
+            if exists(path) and not isdir(path):
                 unlink(path)
                 return json(Response.success('删除成功'))
             else:
@@ -69,9 +79,9 @@ async def download(request: sanic.Request):
     if isinstance(request.json, dict):
         filename = request.json.get('filename', None)
         if isinstance(filename, str):
-            filename = secure_filename(filename)
+            filename = secure_filename_with_directory(filename)
             path = join(system_config.UPLOAD_PATH, filename)
-            if exists(path):
+            if exists(path) and not isdir(path):
                 return await sanic.response.file(path, filename=filename)
             else:
                 return json(Response.failed('文件不存在'), 500)
@@ -87,9 +97,9 @@ async def preview(request: sanic.Request):
     if isinstance(request.json, dict):
         filename = request.json.get('filename', None)
         if isinstance(filename, str):
-            filename = secure_filename(filename)
+            filename = secure_filename_with_directory(filename)
             path = join(system_config.UPLOAD_PATH, filename)
-            if exists(path):
+            if exists(path) and not isdir(path):
                 if getsize(path) < system_config.MAX_PREVIEW_SIZE:
                     content = await read_file(path)
                     try:
@@ -100,7 +110,7 @@ async def preview(request: sanic.Request):
                 else:
                     return json(Response.failed('文件过大, 无法预览'))
             else:
-                return json(Response.failed('文件不存在'))
+                return json(Response.failed('文件不存在或者为文件夹'))
         else:
             return json(Response.invalid('参数无效'))
     else:
@@ -116,19 +126,85 @@ async def modify(request: sanic.Request):
         content = request.json.get('content', None)
 
         if isinstance(filename, str):
-            filename = secure_filename(filename)
+            filename = secure_filename_with_directory(filename)
             path = join(system_config.UPLOAD_PATH, filename)
-            if exists(path):
+            if exists(path) and not isdir(path):
                 if isinstance(content, str):
                     await write_file(path, content.encode())
 
                 if isinstance(new_filename, str):
-                    new_filename = secure_filename(new_filename)
-                    rename(path, join(system_config.UPLOAD_PATH, new_filename))
-
-                return json(Response.success('修改成功'))
+                    new_filename = secure_filename_with_directory(new_filename)
+                    new_path = join(system_config.UPLOAD_PATH, new_filename)
+                    if not exists(new_path):
+                        try:
+                            rename(path, new_path)
+                            return json(Response.success('修改成功'))
+                        except FileNotFoundError:
+                            return json(Response.failed('重命名失败, 请确认新文件所在文件夹存在'))
+                    else:
+                        return json(Response.failed('重命名失败, 目标文件名已存在'))
             else:
                 return json(Response.failed('文件不存在'))
+        else:
+            return json(Response.invalid('参数无效'))
+    else:
+        return json(Response.invalid('参数无效'))
+
+
+@upload_file_controller.route('/add_directory', methods=['POST'])
+@auth_required
+async def add_directory(request: sanic.Request):
+    if isinstance(request.json, dict):
+        directory_name = request.json.get('directory_name', None)
+        if isinstance(directory_name, str) and len(secure_filename(directory_name)) > 0:
+            directory_name = secure_filename(directory_name)
+            full_path = join(system_config.UPLOAD_PATH, directory_name)
+            if not exists(full_path):
+                mkdir(full_path)
+                return json(Response.success('创建成功'))
+            else:
+                return json(Response.failed('文件夹已存在'))
+        else:
+            return json(Response.invalid('参数无效'))
+    else:
+        return json(Response.invalid('参数无效'))
+
+
+@upload_file_controller.route('/delete_directory', methods=['POST'])
+@auth_required
+async def delete_directory(request: sanic.Request):
+    if isinstance(request.json, dict):
+        directory_name = request.json.get('directory_name', None)
+        if isinstance(directory_name, str) and len(secure_filename(directory_name)) > 0:
+            directory_name = secure_filename(directory_name)
+            full_path = join(system_config.UPLOAD_PATH, directory_name)
+            if exists(full_path) and not isfile(full_path):
+                shutil.rmtree(full_path)
+                return json(Response.success('删除成功'))
+            else:
+                return json(Response.failed('文件夹不存在'))
+        else:
+            return json(Response.invalid('参数无效'))
+    else:
+        return json(Response.invalid('参数无效'))
+
+
+@upload_file_controller.route('modify_directory', methods=['POST'])
+@auth_required
+async def modify_directory(request: sanic.Request):
+    if isinstance(request.json, dict):
+        directory_name = request.json.get('directory_name', None)
+        new_directory_name = request.json.get('new_directory_name', None)
+
+        if isinstance(directory_name, str) and len(secure_filename(directory_name)) > 0 and len(secure_filename(new_directory_name)) > 0:
+            full_path = join(system_config.UPLOAD_PATH, secure_filename(directory_name))
+            new_full_path = join(system_config.UPLOAD_PATH, secure_filename(new_directory_name))
+
+            if exists(full_path) and not exists(new_full_path):
+                rename(full_path, new_full_path)
+                return json(Response.success('重命名成功'))
+            else:
+                return json(Response.failed('目标文件夹名已存在'))
         else:
             return json(Response.invalid('参数无效'))
     else:
